added delete function moved to route
